requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.10.6
pydantic-settings==2.7.1
httpx==0.27.2
python-dateutil==2.9.0.post0
PyYAML==6.0.2
SQLAlchemy==2.0.36
asyncpg==0.30.0
psycopg[binary]==3.2.3
alembic==1.14.0
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.5.2
mlflow==2.18.0
matplotlib==3.9.2
plotly==5.24.1
streamlit==1.41.1
structlog==24.4.0
prometheus-client==0.21.1
pytest==8.3.4
pytest-asyncio==0.24.0
joblib==1.4.2

